requirements.txt
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-bootstrap-components==1.5.0
dash-table==5.0.0
Flask==3.0.3
google-auth==2.37.0
google-auth-oauthlib==1.2.1
gspread==6.1.4
httplib2==0.22.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
nest-asyncio==1.6.0
numpy==2.2.1
oauth2client==4.1.3
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
plotly==5.24.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
retrying==1.3.4
rsa==4.9
six==1.17.0
tenacity==9.0.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
Werkzeug==3.0.6
zipp==3.21.0